requirements.txt
# Web Deep Research 依存関係

# ウェブフレームワーク
fastapi>=0.95.0
uvicorn>=0.22.0
flask>=2.2.3

# データ処理
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
nest-asyncio>=1.5.6

# Google AI
google-generativeai>=0.3.0
langchain>=0.0.267
langchain-community>=0.0.10

# ウェブクローリング
selenium>=4.10.0
webdriver-manager>=3.8.6

# グラフ生成
networkx>=3.1

# GUI
PyQt6>=6.5.0
PyQt6-WebEngine>=6.5.0

# モニタリング
prometheus-client>=0.17.0

# テスト
pytest>=7.3.1

# ユーティリティ
tqdm>=4.65.0
colorama>=0.4.6 